SAM-78 Styling fixes
